Rebuilt using latest verion of compose compiler
Updating to clear this error that we are seeing on Jenkins. I believe it is
causing the rise in required heap to prevent an out of memory issue. This was
the reported warning:

Some of the dependencies were build using an older version of the Compose compiler plugin, which may cause additional (or endless) recompositions on non-JVM targets. To prevent that consider updating dependency libraries to versions built with a newer Compose compiler.